+++ src/ska_sdp_piper/piper/utils/__init__.py
@@ -1,12 +1,4 @@
-from .io_utils import (
-    create_output_dir,
-    read_dataset,
-    read_processing_set,
-    read_yml,
-    timestamp,
-    write_dataset,
-    write_yml,
-)
+from .io_utils import create_output_dir, read_yml, timestamp, write_yml
 from .log_util import LogUtil, delayed_log
 
 __all__ = [
@@ -14,9 +6,6 @@ __all__ = [
     "delayed_log",
     "timestamp",
     "LogUtil",
-    "read_dataset",
-    "read_processing_set",
     "read_yml",
-    "write_dataset",
     "write_yml",
 ]