+++ src/ska_sdp_piper/piper/utils/io_utils.py
@@ -3,16 +3,6 @@ from datetime import datetime
 from functools import cache
 
 import yaml
-from xradio.vis.read_processing_set import read_processing_set
-
-
-def read_dataset(infile: str):
-    # Dask related setups.
-    return read_processing_set(ps_store=infile)
-
-
-def write_dataset(output, outfile: str):
-    pass
 
 
 def create_output_dir(output_path, prefix_name):