+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add dask worker plugin to configure logger. Introduce delayed logger
 - Move export image name and format to imaging stage, with the addition of export flags for image, psf, model, and residual.
 - Remove all export stages. The export happens from the respective stages where the data is generated.
+- Move read processing set logic from piper framework to spectral line imaging pipeline
 
 ## 0.5.0